fix(export): moved maxcanvas calculation to export menu to prevent unnecessary heavy calculations
