+++ src/menus/ExportMenu/ExportMenu.js
@@ -1,10 +1,13 @@
-import React, { useState, useRef } from 'react';
+import React, { useState, useRef, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 import { FaDownload } from 'react-icons/fa';
 import { makeStyles } from '@material-ui/core';
 import MenuItem from '../../components/Menu/MenuItem';
 import ExportButton from '../../components/ExportButton';
+import determineMaxCanvasSize from '../../utils/canvasSize';
+
+const LS_SIZE_KEY = 'tm.max.canvas.size';
 
 const useStyles = makeStyles(() => ({
   menuContent: {
@@ -17,11 +20,32 @@ const useStyles = makeStyles(() => ({
 const ExportMenu = () => {
   const classes = useStyles();
   const [collapsed, setCollapsed] = useState(false);
+  const [maxCanvasSize, setMaxCanvasSize] = useState(
+    localStorage.getItem(LS_SIZE_KEY),
+  );
   const permissionInfos = useSelector((state) => state.app.permissionInfos);
   const { t } = useTranslation();
   const ref = useRef();
 
-  if (!permissionInfos || !permissionInfos.user) {
+  useEffect(() => {
+    let timeout = null;
+    if (maxCanvasSize) {
+      return () => {};
+    }
+    timeout = setTimeout(() => {
+      const size = determineMaxCanvasSize();
+      if (size) {
+        setMaxCanvasSize(size);
+        localStorage.setItem(LS_SIZE_KEY, size);
+      }
+    }, 10);
+    return () => {
+      clearTimeout(timeout);
+    };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  if (!permissionInfos || !permissionInfos.user || !maxCanvasSize) {
     return null;
   }
 
@@ -35,14 +59,28 @@ const ExportMenu = () => {
         onCollapseToggle={(c) => setCollapsed(c)}
       >
         <div className={classes.menuContent}>
-          <ExportButton>A0 (72dpi)</ExportButton>
-          <ExportButton exportScale={2}>A0 (150dpi)</ExportButton>
-          <ExportButton exportScale={3}>A0 (300dpi)</ExportButton>
-          <ExportButton exportFormat="a1">A1 (72dpi)</ExportButton>
-          <ExportButton exportFormat="a1" exportScale={2}>
+          <ExportButton maxCanvasSize={maxCanvasSize}>A0 (72dpi)</ExportButton>
+          <ExportButton exportScale={2} maxCanvasSize={maxCanvasSize}>
+            A0 (150dpi)
+          </ExportButton>
+          <ExportButton exportScale={3} maxCanvasSize={maxCanvasSize}>
+            A0 (300dpi)
+          </ExportButton>
+          <ExportButton exportFormat="a1" maxCanvasSize={maxCanvasSize}>
+            A1 (72dpi)
+          </ExportButton>
+          <ExportButton
+            exportFormat="a1"
+            exportScale={2}
+            maxCanvasSize={maxCanvasSize}
+          >
             A1 (150dpi)
           </ExportButton>
-          <ExportButton exportFormat="a1" exportScale={3}>
+          <ExportButton
+            exportFormat="a1"
+            exportScale={3}
+            maxCanvasSize={maxCanvasSize}
+          >
             A1 (300dpi)
           </ExportButton>
         </div>
